requirements.txt
fastapi
uvicorn
pillow
numpy
Tensorflow
requests
python-multipart